refactor(core): split encoding into multiple submodules
